🦄 refactor: shared type.
