+++ apps/ngx-speller-fe/src/features/spelling/spelling-card/spelling-card.component.ts
@@ -2,7 +2,8 @@ import { CommonModule } from '@angular/common';
 import { Component, EventEmitter, HostListener, inject, Input, Output } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { TrustHtmlPipe } from '../../../pipes/trust-html.pipe';
-import { EmitParams, ModeType, WordsItem } from '../../../types';
+import { EmitParams, ModeType } from '../../../types';
+import { WordsItem } from '@shared/types';
 import { BANNED_KEYS, isChineseSymbol } from '../../../utils';
 import { HighlightComponent } from '../../../widgets/highlight/highlight.component';
 import { HornComponent } from '../../../widgets/horn/horn.component';
