+++ apps/ngx-speller-fe/src/store/words/words.actions.ts
@@ -1,6 +1,6 @@
 import { createAction, props } from '@ngrx/store';
 
-import { WordsItem } from '../../types';
+import { WordsItem } from '@shared/types';
 
 export const setWordsList = createAction(
   '[Words Component] Set Words List',
