+++ apps/ngx-speller-fe/src/store/words/words.reducer.ts
@@ -1,6 +1,7 @@
 import { createReducer, on } from '@ngrx/store';
 
-import { WordsItem } from '../../types';
+import { WordsItem } from '@shared/types';
+
 import {
   changeCurrentWordByStep,
   setWordsList,
