+++ apps/ngx-speller-fe/src/types/db.config.type.ts
@@ -1,5 +1,5 @@
 import { Settings } from './settings.type';
-import { WordsItem } from './words.type';
+import { WordsItem } from '@shared/types';
 
 export type WholeIndexDBConfig = {
   settings: Settings;
